requirements.txt
cryptography==2.3.1
ansicolors==1.1.7
